Make config-repos v2 an internal v1 endpoint instead
We will revisit the endpoints later to see which ones need to be public.

Endpoint is available at /go/api/internal/config_repos(/repo_name?)
