+++ .gitignore
@@ -109,8 +109,8 @@ api/api-current-user-v1/target/
 api/api-config-repo-operations-v1/target/
 api/api-config-repo-operations-v1/logs/
 api/api-config-repo-operations-v1/out/
-api/api-config-repos-v2/config/
-api/api-config-repos-v2/out/
+api/api-config-repos-v1/config/
+api/api-config-repos-v1/out/
 api/api-dashboard-v2/logs/
 api/api-dashboard-v2/out/
 api/api-dashboard-v2/target/
