--- api/api-config-repos-v2/src/test/groovy/com/thoughtworks/go/apiv2/configrepos/representers/ConfigReposRepresenterV2Test.groovy
@@ -1,88 +1,0 @@
-/*
- * Copyright 2018 ThoughtWorks, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.thoughtworks.go.apiv2.configrepos.representers
-
-import com.thoughtworks.go.config.PartialConfigParseResult
-import com.thoughtworks.go.config.materials.mercurial.HgMaterialConfig
-import com.thoughtworks.go.config.remote.ConfigRepoConfig
-import com.thoughtworks.go.config.remote.ConfigReposConfig
-import com.thoughtworks.go.config.remote.PartialConfig
-import com.thoughtworks.go.spark.Routes
-import org.junit.jupiter.api.Test
-
-import static com.thoughtworks.go.api.base.JsonUtils.toObjectString
-import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson
-
-class ConfigReposRepresenterV2Test {
-  private static final String TEST_PLUGIN_ID = "test.configrepo.plugin"
-  private static final String TEST_REPO_URL = "https://fakeurl.com"
-
-  @Test
-  void toJSON() {
-    ConfigReposConfig repos = new ConfigReposConfig(repo("foo"), repo("bar"))
-    String json = toObjectString({ w -> ConfigReposRepresenterV2.toJSON(w, repos, resolver()) })
-
-    assertThatJson(json).isEqualTo([
-      _links   : [
-        self: [href: "http://test.host/go$Routes.ConfigRepos.BASE".toString()]
-      ],
-      _embedded: [
-        config_repos: [
-          expectedRepoJson("foo"),
-          expectedRepoJson("bar")
-        ]
-      ]
-    ])
-  }
-
-  static Map expectedRepoJson(String id) {
-    return [
-      _links       : [
-        self: [href: "http://test.host/go${Routes.ConfigRepos.id(id)}".toString()],
-        doc : [href: Routes.ConfigRepos.DOC],
-        find: [href: "http://test.host/go${Routes.ConfigRepos.find()}".toString()],
-      ],
-
-      id           : id,
-      plugin_id    : TEST_PLUGIN_ID,
-      material     : [
-        type      : "hg",
-        attributes: [
-          name       : null,
-          url        : "${TEST_REPO_URL}/$id".toString(),
-          auto_update: true
-        ]
-      ],
-      configuration: [],
-
-      last_parse   : [
-        revision: "$id-123".toString(),
-        success : true,
-        error   : null
-      ]
-    ]
-  }
-
-  static ConfigRepoConfig repo(String id) {
-    HgMaterialConfig materialConfig = new HgMaterialConfig("$TEST_REPO_URL/$id", "")
-    return new ConfigRepoConfig(materialConfig, TEST_PLUGIN_ID, id)
-  }
-
-  static ConfigRepoResultResolver resolver() {
-    return { r -> new PartialConfigParseResult("${r.getId()}-123", new PartialConfig()) }
-  }
-}
