--- api/api-config-repos-v2/src/test/groovy/com/thoughtworks/go/apiv2/configrepos/representers/MaterialRepresenterTest.groovy
@@ -1,60 +1,0 @@
-/*
- * Copyright 2018 ThoughtWorks, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.thoughtworks.go.apiv2.configrepos.representers
-
-import com.thoughtworks.go.api.base.OutputWriter
-import com.thoughtworks.go.config.materials.PackageMaterialConfig
-import com.thoughtworks.go.config.materials.git.GitMaterialConfig
-import com.thoughtworks.go.domain.materials.MaterialConfig
-import org.junit.jupiter.api.Test
-
-import static com.thoughtworks.go.api.base.JsonUtils.toObjectString
-import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson
-import static org.junit.jupiter.api.Assertions.assertEquals
-import static org.junit.jupiter.api.Assertions.assertThrows
-import static org.mockito.Mockito.mock
-
-class MaterialRepresenterTest {
-  private static final String REPO_URL = "https://guthib.com/chewbacca"
-  private static final String BRANCH = "wookie"
-
-  @Test
-  void toJSON() {
-    MaterialConfig config = new GitMaterialConfig(REPO_URL, BRANCH)
-    String json = toObjectString({ w -> MaterialRepresenter.toJSON(w, config) })
-
-    assertThatJson(json).isEqualTo([
-      type      : "git",
-      attributes: [
-        name       : null,
-        url        : REPO_URL,
-        branch     : BRANCH,
-        auto_update: true
-      ]
-    ])
-  }
-
-  @Test
-  void 'toJSON() fails on unsupported material type'() {
-    MaterialConfig config = new PackageMaterialConfig("foobar");
-    Throwable error = assertThrows(IllegalArgumentException.class, {
-      MaterialRepresenter.toJSON(mock(OutputWriter.class), config)
-    })
-
-    assertEquals("Cannot serialize unsupported material type: PackageMaterialConfig", error.getMessage())
-  }
-}
