+++ server/webapp/WEB-INF/applicationContext-global.xml
@@ -27,7 +27,7 @@
 
   <context:annotation-config/>
 
-  <context:component-scan base-package="com.thoughtworks.go.apiv2.configrepos"/>
+  <context:component-scan base-package="com.thoughtworks.go.apiv1.configrepos"/>
   <context:component-scan base-package="com.thoughtworks.go.apiv1.configrepooperations"/>
   <context:component-scan base-package="com.thoughtworks.go.apiv1.elasticprofile"/>
   <context:component-scan base-package="com.thoughtworks.go.apiv1.admin.security"/>
